requirements.txt
flask>=2.0.0
fpdf2>=2.8.0
